+++ MSVCRTTI/CompleteObjectLocator.cs
@@ -4,21 +4,21 @@ using System.Runtime.InteropServices;
 
 namespace Henke37.DebugHelp.RTTI.MSVC {
 	internal class CompleteObjectLocator {
-		public uint ObjectRootOffset;
-		public uint ConstructorDescriptorOffset;
+		public int ObjectRootOffset;
+		public int ConstructorDescriptorOffset;
 		public TypeDescriptor TypeDescriptor;
 		public ClassHierarchyDescriptor ClassHierarchyDescriptor;
 
 		[StructLayout(LayoutKind.Sequential, Pack = 1)]
 		internal struct MemoryStruct {
 			public uint Signature;
-			public uint ObjectRootOffset;
-			public uint ConstructorDescriptorOffset;
-			public uint pTypeDescriptor;
-			public uint pClassDescriptor;
+			public int ObjectRootOffset;
+			public int ConstructorDescriptorOffset;
+			public IntPtr pTypeDescriptor;
+			public IntPtr pClassDescriptor;
 		}
 
-		public uint LocateCompleteObject(uint objAddr) {
+		public IntPtr LocateCompleteObject(IntPtr objAddr) {
 			return objAddr - ObjectRootOffset;
 		}
 	}
